activity controller is done, need to figure out bettwe redirects for post and put requests working on comments controller next

+++ src/main/java/com/groupeight/citiesandactivities/controllers/HomeController.java
@@ -34,7 +34,7 @@ public class HomeController {
 		User user = userService.findByUsername(username);
 
 		if (session.getAttribute("userId") == null) {
-			Integer thisUsersId = user.getId();
+			Long thisUsersId = user.getId();
 			session.setAttribute("userId", thisUsersId);
 
 		}
